requirements.txt
streamlit
ollama
requests
langchain
langchain-community
langchain-text-splitters
pymupdf
sentence-transformers
mysql-connector-python
beautifulsoup4
python-dotenv
